Ajout de la securisation des access

+++ src/Controller/TaskController.php
@@ -13,21 +13,18 @@ use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
+#[IsGranted('ROLE_USER')]
 class TaskController extends AbstractController
 {
-    /**
-     * @Route("/tasks", name="task_list")
-     */
+    #[Route('/tasks', name: 'task_list')]
     public function listAction(TaskRepository $taskRepository): Response
     {
-        return $this->render('task/list.html.twig', ['tasks' => $taskRepository->findAll()]);
+        return $this->render('task/list.html.twig', ['tasks' => $taskRepository->findBy(['user' => $this->getUser()])]);
     }
 
-    /**
-     * @Route("/tasks/create", name="task_create")
-     * @throws ORMException
-     */
+    #[Route('/tasks/create', name: 'task_create')]
     public function createAction(Request $request, EntityManagerInterface $em): Response
     {
         $task = new Task();
@@ -47,21 +44,16 @@ class TaskController extends AbstractController
         return $this->render('task/create.html.twig', ['form' => $form]);
     }
 
-    /**
-     * @Route("/tasks/{id}/edit", name="task_edit")
-     */
+    #[Route('/tasks/{id}/edit', name: 'task_edit')]
     public function editAction(Task $task, Request $request, EntityManagerInterface $em): RedirectResponse|Response
     {
+        $this->denyAccessUnlessGranted('TASK_EDIT', $task);
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            try {
-                $em->flush();
-            } catch (OptimisticLockException|ORMException $e) {
-            }
-
+            $em->flush();
             $this->addFlash('success', 'La tâche a bien été modifiée.');
 
             return $this->redirectToRoute('task_list');
@@ -72,36 +64,29 @@ class TaskController extends AbstractController
             'task' => $task,
         ]);
     }
-    /**
-     * @Route("/tasks/done", name="task_list_done")
-     */
+
+    #[Route('/tasks/done', name: 'task_list_done')]
     public function listDone(TaskRepository $taskRepository): Response
     {
         return $this->render('task/is_done.html.twig', ['tasks' => $taskRepository->findBy(['isDone' => true])]);
     }
 
-    /**
-     * @Route("/tasks/{id}/toggle", name="task_toggle")
-     */
+    #[Route('/tasks/{id}/toggle', name: 'task_toggle')]
     public function toggleTaskAction(Task $task, EntityManagerInterface $em): RedirectResponse
     {
         $task->toggle(!$task->isDone());
-        try {
-            $em->flush();
-        } catch (OptimisticLockException|ORMException $e) {
-        }
+        $em->flush();
 
         $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
 
         return $this->redirectToRoute('task_list');
     }
 
-    /**
-     * @Route("/tasks/{id}/delete", name="task_delete")
-     * @throws ORMException
-     */
+    #[Route('/tasks/{id}/delete', name: 'task_delete')]
     public function deleteTaskAction(Task $task, EntityManagerInterface $em): RedirectResponse
     {
+
+        $this->denyAccessUnlessGranted('TASK_DELETE', $task);
         $em->remove($task);
         $em->flush();
